Here are some examples of the Remote Linux Lab
These examples are of disabling power management for the Remote Linux Lab, since we don't want people powering off the server by accident.  The disable-colorpolicy is there to keep gnome3 from asking for a admin password or root when a user first logs in.  I would put more .dot files, but most of the time like for vim, if i need to see the line numbers I look at the bottom of the screen and if I really need to see them there's always :set number.